rabbitmq-tracer/requirements.txt
aio-pika==9.5.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
